+++ js/eve-online-intel-tool.js
@@ -129,8 +129,17 @@ jQuery(document).ready(function($) {
      * todo:
      * Get this to work with data tables pagination, so it's deactivated for now
      */
-    $('tr[data-highlight]').click(function() {
+    $('tr[data-highlight]').on('click', function(e) {
+//        if (e.target !== this) {
+//            console.log(e.target);
+//        }
+
+        console.log($(this).attr("class"));
+
         $('tr[data-highlight="' + $(this).data('highlight') + '"]').toggleClass('dataHighlightSticky');
+    }).on('click', '.eve-intel-information-link', function(e) {
+        // clicked on descendant div
+        e.stopPropagation();
     });
 
     /**