+++ templates/partials/fleet/fleet-composition.php
@@ -22,9 +22,16 @@ if(\is_array($fleetOverview) && \count($fleetOverview)) {
                             ],
                             'pluginSettings' => $pluginSettings
                         ]);
-
-                        echo $data['pilotName'];
                         ?>
+
+                        <span class="eve-intel-pilot-information-wrapper">
+                            <span class="eve-intel-pilot-name-wrapper">
+                                <?php echo $data['pilotName']; ?>
+                            </span>
+                            <span class="eve-intel-pilot-links-wrapper">
+                                <small><a class="eve-intel-information-link" href="https://evewho.com/pilot/<?php echo \urlencode($data['pilotName']); ?>" target="_blank">evewho <i class="fa fa-external-link" aria-hidden="true"></i></a> | <a class="eve-intel-information-link" href="https://zkillboard.com/character/<?php echo $data['pilotName']; ?>/" target="_blank">zkillboard <i class="fa fa-external-link" aria-hidden="true"></i></a></small>
+                            </span>
+                        </span>
                     </td>
                     <td>
                         <?php echo $data['shipClass']; ?>