+++ templates/partials/pilot/pilot-participation.php
@@ -22,9 +22,16 @@ if(!empty($pilotParticipation)) {
                             'data' => $pilot,
                             'pluginSettings' => $pluginSettings
                         ]);
-
-                        echo $pilot['characterName'];
                         ?>
+
+                        <span class="eve-intel-pilot-information-wrapper">
+                            <span class="eve-intel-pilot-name-wrapper">
+                                <?php echo $pilot['characterName']; ?>
+                            </span>
+                            <span class="eve-intel-pilot-links-wrapper">
+                                <small><a class="eve-intel-information-link" href="https://evewho.com/pilot/<?php echo \urlencode($pilot['characterName']); ?>" target="_blank">evewho <i class="fa fa-external-link" aria-hidden="true"></i></a> | <a class="eve-intel-information-link" href="https://zkillboard.com/character/<?php echo $pilot['characterName']; ?>/" target="_blank">zkillboard <i class="fa fa-external-link" aria-hidden="true"></i></a></small>
+                            </span>
+                        </span>
                     </td>
 
                     <td>